[Chore] Move components into page folder
