+++ src/app/about/components/SocialSection.tsx
@@ -2,7 +2,7 @@ import { Github, Linkedin, Mail } from 'lucide-react';
 import Link from 'next/link';
 import type { ReactNode } from 'react';
 
-import { Button } from './ui/button';
+import { Button } from '../../../components/ui/button';
 
 import { cn } from '@/lib/utils';
